Move action to the new namespace
